Katalyst-TDD-Starter/Katalyst-TDD-Starter/Bags/BagBelt.cs
using Katalyst_TDD_Starter.Test.Bags;

namespace Katalyst_TDD_Starter.Bags
{
    public class BagBelt
    {
        public List<StorageBag> Bags { get; set; } = new();

        public void AddBag(StorageBag input)
        {
            Bags.Add(input);
        }

        public void AddItem(Item item)
        {
            var openCategoryBags = Bags.Where(x => x.Category == item.Category && x.HasSpace()).ToList();

            if (openCategoryBags.Any())
            {
                openCategoryBags.First().Add(item);
                return;
            }

            var openBags = Bags.Where(x => x.HasSpace()).ToList();
            if (openBags.Any())
            {
                openBags.First().Add(item);
                return;
            }
        }
    }
}